Got everything working
This adds a basic CLI that can execute anything that systemconf can
describe and display it in a table for both checks and installs. Next up
is gong to be refactoring the process execution to show the current line
from the subprocess on the active row so you can tell what's going on
when you're waiting. Thta's going to take a bit of refactoring. I'll
probably need to add an actual Executble class to hold some of the logic
that I keep having to fork between recipes invocations and shell lines.