Fix display bug while setting up in table
The table used to leave behind artifacts during the setup process. I
think this happened because the table would grow and then shrink in
height, leaving behind w/e was at the top most portion of the table.
This change adds some dynamic padding to the table so it never shrinks
in height, but also doesn't pre-reserve the max padding in case none of
the targets cause the table to grow.