Make table updates dynamic
This adds a glorified dirty-flag checker to make the table updates
dynamic to the amount of output. I noticed how many refreshes were
happening when I tried to record the asciinema demo and got a 40mb file
out of it. Now there are min/max update constraints so the refresh rate
slows when nothing is happening but speeds up when there is a lot of
output from one of the subprocesses.